Fix Card and Deck reading from Parcel.
‘Id’ numbers should be read as ‘long’ values, not ‘int’ to respect a171565bff1.

+++ src/app/android/gambit/local/Card.java
@@ -63,7 +63,7 @@ public class Card implements Parcelable
 	}
 
 	public void readFromParcel(Parcel parcel) {
-		id = parcel.readInt();
+		id = parcel.readLong();
 		frontSideText = parcel.readString();
 		backSideText = parcel.readString();
 	}
